moved layout component
