+++ src/components/Blog/Post/index.tsx
@@ -1,6 +1,6 @@
 import React from "react"
-import Layout from "../../Layout"
-import TitleGrid from "../../layout/TitleGrid"
+import Layout from "../../specific/Layout"
+import TitleGrid from "../../specific/layout/TitleGrid"
 import SEO from "../SEO"
 import Img from "gatsby-image"
 import styles from "./styles.module.scss"
