+++ src/pages/404.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 
-import Layout from "../components/Layout"
+import Layout from "../components/specific/Layout"
 import SEO from "../components/Blog/SEO"
 
 const NotFoundPage: React.FC = () => (
