+++ src/pages/index.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import Layout from "../components/Layout"
+import Layout from "../components/specific/Layout"
 import SEO from "../components/Blog/SEO"
 
 const IndexPage: React.FC = () => (
